+++ src/graphology/tests/utils.js
@@ -182,45 +182,6 @@ export default function utils(Graph) {
 
         assert.deepStrictEqual(graph.getAttributes(), copy.getAttributes());
       }
-    },
-
-    '#.upgradeToMixed': {
-      'it should correctly upgrade the given graph to a mixed one.':
-        function () {
-          const graph = new Graph({type: 'directed'});
-          addNodesFrom(graph, [1, 2, 3]);
-          graph.addEdgeWithKey('1->2', 1, 2);
-
-          graph.upgradeToMixed();
-
-          assert.strictEqual(graph.type, 'mixed');
-
-          graph.addUndirectedEdge(2, 3);
-
-          assert.strictEqual(graph.size, 2);
-          assert.strictEqual(graph.degree(2), 2);
-        }
-    },
-
-    '#.upgradeToMulti': {
-      'it should correctly upgrade the given graph to a multi one.':
-        function () {
-          const graph = new Graph();
-          addNodesFrom(graph, [1, 2]);
-          graph.addEdgeWithKey('dA', 1, 2);
-          graph.addUndirectedEdgeWithKey('uA', 1, 2);
-
-          graph.upgradeToMulti();
-
-          assert.strictEqual(graph.multi, true);
-
-          graph.addEdgeWithKey('dB', 1, 2);
-          graph.addUndirectedEdgeWithKey('uB', 1, 2);
-
-          assert.strictEqual(graph.size, 4);
-
-          assert.deepStrictEqual(graph.edges(), ['dA', 'uA', 'dB', 'uB']);
-        }
     }
   };
 }