+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.23.3 (provisional)
 
 - Improving performance of internal edge key generator.
+- Refactoring internal edge & neighbor iteration schemes.
 
 ## 0.23.2